workspace: refactored into new daemon/service architecture
